fewuo

+++ PingClient.c
@@ -90,15 +90,29 @@ int main(int argc, char *argv[]) {
         sleep(1);
     }
 
-    double loss = ( (double) (sent - received) / (double)sent) * 100.0;
-    double avg_rtt = (received > 0) ? (sum_rtt / (double) received) : 0.0;
+    // double loss = ( (double) (sent - received) / (double)sent) * 100.0;
+    // double avg_rtt = (received > 0) ? (sum_rtt / (double) received) : 0.0;
+
+    // printf("--- %s ping statistics ---\n", server_name);
+    // printf("%d packets transmitted, %d received, %d%% packet loss ",
+    //        sent, received, (int)loss);
+    // if (received > 0) {
+    //     printf("rtt min/avg/max = %.3f %.3f %.3f ms\n",
+    //            min_rtt, avg_rtt, max_rtt);
+    // }
+
+    // close(sockfd);
+    // return 0;
+    double loss = ((double)(sent - received) / (double)sent) * 100.0;
+    double avg_rtt = (received > 0) ? (sum_rtt / (double)received) : 0.0;
 
     printf("--- %s ping statistics ---\n", server_name);
-    printf("%d packets transmitted, %d received, %d%% packet loss\n",
-           sent, received, (int)loss);
+    printf("%d packets transmitted, %d received, %.0f%% packet loss\n",
+        sent, received, loss);
+
     if (received > 0) {
         printf("rtt min/avg/max = %.3f %.3f %.3f ms\n",
-               min_rtt, avg_rtt, max_rtt);
+            min_rtt, avg_rtt, max_rtt);
     }
 
     close(sockfd);
